+++ backend/utils/socket.js
@@ -1,0 +1,54 @@
+const socketIO=require("socket.io");
+
+let users = [];
+
+const addUser = (userId, socketId) => {
+  !users.some((user) => user.userId === userId) &&
+    users.push({ userId, socketId });
+};
+
+const removeUser = (socketId) => {
+  users = users.filter((user) => user.socketId !== socketId);
+};
+
+const getUser = (userId) => {
+  return users.find((user) => user.userId === userId);
+};
+
+const initializeSocket = (server) => {
+    const io = socketIO(server, {
+      cors: {
+        origin: "*",
+      },
+    });
+
+io.on("connection", (socket) => {
+  console.log("a user connected.");
+
+  // connection
+  socket.on("addUser", (userId) => {
+    addUser(userId, socket.id);
+    io.emit("getUsers", users);
+  });
+
+  // send and receive messages
+  socket.on("sendMessage", ({ senderId, receiverId, text }) => {
+    const user = getUser(receiverId);
+    io.to(user?.socketId).emit("getMessage", {
+      senderId,
+      text,
+    });
+  });
+
+  // disconnection
+  socket.on("disconnect", () => {
+    console.log("a user disconnected!");
+    removeUser(socket.id);
+    io.emit("getUsers", users);
+  });
+});
+
+return io;
+}
+
+module.exports=initializeSocket;
